increment1 Перенес код агента в internal пакет
